WIP: test-drive atomics
See #126

Initial performance tests with atomics are unimpressive. Will only
be noticeable if either a bunch of debits are coming from one
account or a bunch of credits are going to one account. The
current demo creates a bunch of different accounts, and the
current stage benchmark uses all orthogonal transactions.