WIP: Parallelize contract language
This unfortunately won't work as-is unless the leader can first
ensure that no two events in the same batch affect the same contract
in non-deterministic ways. For signatures this is straightforward,
but for timestamps we'd need to check every every contract to see
if its waiting on a timestamp. For example, if there was a race
between a signature and a timestamp (or timestamps at two different
times), then the contract would take a different route depending
on which it saw first.